Moving Settings fields to Settings_Handler
